requirements-airflow.txt
apache-airflow-providers-postgres==5.6.0
pandas==2.0.3
sqlalchemy<2.0
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-mock==3.10.0
pytest-cov==4.1.0
numpy==1.24.3
requests==2.31.0 